backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.9.2
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
apscheduler==3.10.4
python-multipart==0.0.12
firebase-admin==6.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
